feat: Treat withState actions as side effects
Fixes #29. This change changes the treatment of GetStateActions.
These actions are now treated as side effects, and the state processor
does not wait for their completion. Instead, it spins off a separate
to execute them, and moves on to process other actions in the queue
without waiting for it to finish.